Created base template of a recipe, and an image folder for recipes
Created an images folder within the recipes folder,
to store the various images of our recipes.

Layed out the basic structure of a recipe page, built first recipe as a
'grilled cheese', now can just copy and paste the html and replace and
add whatever depending on the recipes ingredients and steps.